requirements.txt
# DASK+ Parametrik Sigorta - Backend Requirements
# ================================================

# Web Framework
flask==3.0.0
flask-cors==4.0.0

# HTTP Requests (Kandilli Entegrasyonu)
requests==2.31.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Machine Learning
scikit-learn==1.3.2
scipy==1.11.4

# Geospatial
geopy==2.4.1
pyproj==3.6.1

# Visualization
matplotlib==3.8.2
seaborn==0.13.0
folium==0.15.1

# Utilities
tqdm==4.66.1
python-dateutil==2.8.2

# Web Server (Production)
gunicorn==21.2.0

# Additional
openpyxl==3.1.2

# Security & Cryptography
pycryptodome==3.19.0
pyjwt==2.8.0